[YITEAM-111] 북마크 API BucketController에 합침
